+++ main.py
@@ -4,6 +4,7 @@ from dotenv import load_dotenv
 import os
 import asyncio
 from database.roleta_db import DBroleta
+from utils.embed import  EmbedPadrao
 
 load_dotenv()
 token = os.getenv("TOKEN")
@@ -14,7 +15,8 @@ intents.members = True
 
 bot = commands.Bot(command_prefix="$",
                     intents=intents,
-                    case_insensitive=True)
+                    case_insensitive=True,
+                    help_command=None)
 
 
 
@@ -23,6 +25,17 @@ async def on_ready():
     DBroleta.criar_tabela()
     print("Bot online com sucesso!")
 
+@bot.event
+async def on_command_error(ctx,error):
+    from discord.ext import commands
+
+    if isinstance(error,commands.CommandNotFound):
+        await ctx.send(embed=EmbedPadrao.erro(
+            ctx,
+            "Comando não encontrado, Use $ajuda para ver os comandos."
+        ))
+    else:
+        raise error
 
 async def main():
     async with bot: